+++ changelog.md
@@ -1,10 +1,16 @@
 # Changelog
 
-## v0.2.0
+## v0.3.0 (2016-12-16)
+
+- Subscription commands simplified (`subscribe`, `unsubscribe` and `subscriptions`)
+- Removed irc disconnect notifications (this time for real)
+- Improved JmsClient DSL
+
+## v0.2.0 (2016-11-18)
 
 - Code improvements
-- Add a way of reconnecting to the irc server in case of a connection error
-- Remove irc disconnect notifications
+- Added a way of reconnecting to the irc server in case of a connection error
+- Removed irc disconnect notifications
 
 ## v0.1.0 (2016-11-13)